+++ changes.md
@@ -1,3 +1,11 @@
-### Sept 13 2022
+
+
+### Sept15
+- renamed KEY ennvironment variable to WP_AUTH_TOKEN
+- Added Eslint config.
+- Server now listens after successful DB connection
+
+
+### Sept13
 - Moved index.js into src folder.
 - Added additional build scripts in package.json for Heroku deployment